Fixed a bug

+++ Arduino/V0p2_Main/Control.cpp
@@ -404,7 +404,8 @@ DEBUG_SERIAL_PRINTLN_FLASHSTRING("JSON gen err!");
 #if defined(ENABLE_OTSECUREFRAME_ENCODING_SUPPORT)
       // Explicit-workspace version of encryption.
       const OTRadioLink::SimpleSecureFrame32or0BodyTXBase::fixed32BTextSize12BNonce16BTagSimpleEncWithLWorkspace_ptr_t eW = OTAESGCM::fixed32BTextSize12BNonce16BTagSimpleEnc_DEFAULT_WITH_LWORKSPACE;
-      constexpr uint8_t workspaceSize = OTRadioLink::SimpleSecureFrame32or0BodyTXBase::generateSecureOFrameRawForTX_total_scratch_usage_OTAESGCM_2p0;
+
+      constexpr size_t workspaceSize = OTRadioLink::SimpleSecureFrame32or0BodyTXBase::generateSecureOFrameRawForTX_total_scratch_usage_OTAESGCM_2p0;
       uint8_t workspace[workspaceSize];
       OTV0P2BASE::ScratchSpaceL sW(workspace, workspaceSize);
       const uint8_t txIDLen = OTRadioLink::ENC_BODY_DEFAULT_ID_BYTES;
